:truck: chatgpt api 호출 로직 상위 컴포넌트(ChatbotPage)로 이동
